+++ src/main/java/com/nhnacademy/authservice/client/member/OAuth2MemberClient.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.authservice.client.member;
 
-import com.nhnacademy.authservice.domain.response.OAuth2MemberResponse;
+import com.nhnacademy.authservice.dto.oauth2.response.OAuth2MemberResponse;
 
 public interface OAuth2MemberClient {
     OAuth2MemberResponse getMember(String accessToken);
