+++ src/main/java/com/nhnacademy/authservice/dto/auth/response/TokenParseResponseDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.authservice.domain.response;
+package com.nhnacademy.authservice.dto.auth.response;
 
 import com.nhnacademy.authservice.provider.UserType;
